implicit-casts:false in flutter/lib/src/physics (#45622)

+++ packages/flutter/lib/src/physics/clamped_simulation.dart
@@ -48,10 +48,10 @@ class ClampedSimulation extends Simulation {
   final double dxMax;
 
   @override
-  double x(double time) => simulation.x(time).clamp(xMin, xMax);
+  double x(double time) => simulation.x(time).clamp(xMin, xMax) as double;
 
   @override
-  double dx(double time) => simulation.dx(time).clamp(dxMin, dxMax);
+  double dx(double time) => simulation.dx(time).clamp(dxMin, dxMax) as double;
 
   @override
   bool isDone(double time) => simulation.isDone(time);
